+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 Change Log
 ==========
 
+<a name="v0.0.98"></a>
+## Version [v0.0.98](https://github.com/luizfp/PrologWebService/compare/v0.0.97...v0.0.98) (release-date) [unreleased]
+
+
+
 <a name="v0.0.97"></a>
-## Version [v0.0.97](https://github.com/luizfp/PrologWebService/compare/v0.0.96...v0.0.97) (release-date) [unreleased]
+## Version [v0.0.97](https://github.com/luizfp/PrologWebService/compare/v0.0.96...v0.0.97) (2019-06-28)
 
 ### Features
 * Cria funcionalidade de listagem de marcações otimizada, com filtro de data. (PL-2000)
+* Cria método de busca das regionais e unidades de filtro para início de checklist (PL-2085)
+
+### Refactors
+* Altera método de login para incluir se empresa está bloqueada para realizar checklist de diferentes unidades no `LoginHolder` (PL-2077)
+* Altera método de busca de um checklist por código (PL-2075)
 
 ### Bug Fixes
 * Corrige problema no fechamento de serviço de movimentação (PL-2114)